feat(react): #WB2-1905, add modals as a submodule
