+++ packages/react/ui/src/index.ts
@@ -1,4 +1,3 @@
-export * from './common';
 export * from './components';
 export * from './core';
 export * from './hooks';
