+++ packages/react/ui/src/modules/modals/OnboardingModal/OnboardingModal.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import { Pagination } from 'swiper/modules';
 import { Swiper, SwiperSlide } from 'swiper/react';
 
-import { Button, Image, Modal, usePaths } from '../..';
+import { Button, Image, Modal, usePaths } from '../../..';
 import { useOnboardingModal } from './useOnboardingModal';
 
 interface ModalItemsProps {
