+++ packages/react/ui/src/modules/modals/PublishModal/hooks/usePublishModal.tsx
@@ -5,11 +5,11 @@ import {
   odeServices,
   type PublishParameters,
 } from 'edifice-ts-client';
-import { type SubmitHandler, useForm } from 'react-hook-form';
+import { useForm, type SubmitHandler } from 'react-hook-form';
 
-import { useOdeClient } from '../../../core';
-import { useToast } from '../../../hooks';
-import { libraryMaps } from '../../../utils/libraryMaps';
+import { useOdeClient } from '../../../../core';
+import { useToast } from '../../../../hooks';
+import { libraryMaps } from '../../../../utils/libraryMaps';
 import { ToastError } from '../components/ToastError';
 import { ToastSuccess } from '../components/ToastSuccess';
