+++ packages/react/ui/src/modules/modals/ResourceModal/ResourceModal.tsx
@@ -23,13 +23,13 @@ import {
   LoadingScreen,
   Modal,
   TextArea,
-} from '../../components';
-import { TextareaCounter } from '../../components/TextArea/TextareaCounter';
-import { useMediaLibrary, useOdeClient } from '../../core';
-import { useResource } from '../../core/useResource';
-import { useToast } from '../../hooks';
-import { MediaLibrary } from '../../modules/multimedia';
-import ImagePickerWorkspace from '../../modules/multimedia/ImagePickerWorkspace/ImagePickerWorkspace';
+} from '../../../components';
+import { TextareaCounter } from '../../../components/TextArea/TextareaCounter';
+import { useMediaLibrary, useOdeClient } from '../../../core';
+import { useResource } from '../../../core/useResource';
+import { useToast } from '../../../hooks';
+import { MediaLibrary } from '../../multimedia';
+import ImagePickerWorkspace from '../../multimedia/ImagePickerWorkspace/ImagePickerWorkspace';
 import { useThumb } from './hooks/useThumb';
 
 export interface FormInputs {
