+++ packages/react/ui/src/modules/modals/ResourceModal/hooks/useThumb.ts
