+++ packages/react/ui/src/modules/modals/ResourceModal/index.ts
