+++ packages/react/ui/src/modules/modals/ShareModal/utils/hasRight.ts
