+++ packages/react/ui/src/modules/modals/ShareModal/utils/showShareRightLine.ts
